+++ src/onegov.winterthur/onegov/winterthur/assets/js/iframe-resizer-options.js
@@ -1,0 +1,5 @@
+window.iFrameResizer = {
+    heightCalculationMethod: function() {
+        return $('.content').height();
+    }
+};